added a simple confirmation popup before uninstallation

+++ src/PTGUI/MainWindow.cs
@@ -1,22 +1,20 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
-using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Interactivity;
 using Avalonia.VisualTree;
+using PTG;
 
 namespace PTGUI;
 
-public class MainWindow : Window
+public class MainWindow : BaseWindow
 {
     private Panel? deviceInfoPanel;
     private ListBox? packagesList;
     public MainWindow()
     {
-        Title = "hiiii";
-        MinHeight = 600;
+        Title = "PTGUI";
+        MinHeight = 700;
         MinWidth = 1000;
         InitializeUi();
     }
@@ -56,12 +54,12 @@ public class MainWindow : Window
         sideBar.Children.Add(deviceInfoPanel);
 
 
-        var mainContenet = new StackPanel()
+        var mainContent = new StackPanel()
         {
             Name = "mainContenet",
             [Grid.ColumnProperty] = 1
         };
-        grid.Children.Add(mainContenet);
+        grid.Children.Add(mainContent);
 
         packagesList = new ListBox()
         {
@@ -69,27 +67,33 @@ public class MainWindow : Window
             SelectionMode = SelectionMode.Multiple,
             MaxHeight = 600
         };
-        mainContenet.Children.Add(packagesList);
+        mainContent.Children.Add(packagesList);
 
         var uninstallButton = new Button()
         {
             Content = "uninstall",
         };
-        mainContenet.Children.Add(uninstallButton);
+        mainContent.Children.Add(uninstallButton);
         uninstallButton.Click += uninstallButton_Clicked;
     }
 
-    private void uninstallButton_Clicked(object? sender, RoutedEventArgs e)
+    private async void uninstallButton_Clicked(object? sender, RoutedEventArgs e)
     {
         Console.WriteLine("uninstalling...");
-
         var children = ((Grid)this.Content).GetVisualChildren();
         var deviceList = FindControl<ComboBox>(children, "devicesList");
+        // var deviceList = this.FindNameScope().Find<ComboBox>("devicesList");
 
         if (deviceList is null || deviceList.SelectedItem is null) return;
         if (packagesList is null || packagesList.SelectedItems.Count <= 0) return;
 
         var packages = packagesList.SelectedItems.Cast<string>();
+
+        PopupWindow confirmationPopup = new(packages);
+        await confirmationPopup.ShowDialog(this);
+
+        if(!confirmationPopup.isConfirmed) return;
+
         PlatformTools.Adb.uninstallPackages((string)deviceList.SelectedItem, packages);
     }
 
@@ -126,22 +130,5 @@ public class MainWindow : Window
             text.Text = text.Text + "\n " + info;
         }
         deviceInfoPanel!.Children.Add(text);
-    }
-
-    private T? FindControl<T>(IEnumerable<IVisual> parent, string name) where T : class
-    {
-        foreach (IVisual? child in parent)
-        {
-            if (child.GetType() == typeof(T) && ((StyledElement)child).Name == name)
-                return (T)child;
-            // https://stackoverflow.com/a/10718451/12429279
-            if (typeof(Panel).GetTypeInfo().IsAssignableFrom(child.GetType()))
-            {
-                var x = FindControl<T>(((Panel)child).GetVisualChildren(), name);
-                if (x is not null)
-                    return x;
-            }
-        }
-        return null;
     }
 }
